refactor: Move message component struct & function to it's own file
